pyproject.toml
[project]
name = "update-all-forks"
version = "0.1.0"
description = "Upate all Github forks"
authors = [
    {name = "Myridia Company"}
]
license = {text = "GNU"}
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "requests (>=2.32.5,<3.0.0)",
    "pygithub (>=2.8.1,<3.0.0)",
    "toml (>=0.10.2,<0.11.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false